refactor: move bank transactions state into providers/bankTransactions
Nine concerns move into the domain: the transactions data provider, pagination,
filters, string overrides, feature visibility, route state, bulk selection,
the bank-account filter, and receipts. Each context lands in the same
directory as its provider, so a concern is one directory rather than a pair
split across two trees.

This corrects a swap in the process. providers/BankTransactionsProvider held
BankTransactionsPaginationProvider.tsx and providers/BankTransactionsPagination-
Provider held BankTransactionsProvider.tsx -- each directory contained the
other one's component. The files now sit in directories named for what they
export: bankTransactions/BankTransactions and
bankTransactions/BankTransactionsPagination.

Two placements are judgment calls. BankAccountsFilterStore goes here rather
than under linkedAccounts because the filter narrows the transactions list,
even though features/linkedAccounts reads it too. BulkSelectionStore has a
domain-agnostic API (useSelectedIds, useIdIsSelected), but every consumer
today is a bank-transactions component or blocks/BulkActionsModule; promote it
to global/ when a second domain needs it.
