+++ src/components/features/bankTransactions/BankTransactionFormFields/BankTransactionFormFields.tsx
@@ -4,7 +4,7 @@ import type { BankTransaction } from '@internal-types/bankTransactions'
 import { makeTagFromTransactionTag, type Tag } from '@schemas/tag'
 import { useDeleteBankTransactionTag } from '@api/businesses/[business-id]/bank-transactions/tags/delete'
 import { usePostBankTransactionTag } from '@api/businesses/[business-id]/bank-transactions/tags/post'
-import { BankTransactionsFeature, useIsBankTransactionsFeatureEnabled } from '@providers/BankTransactionsFeatureVisibility/BankTransactionsFeatureVisibilityProvider'
+import { BankTransactionsFeature, useIsBankTransactionsFeatureEnabled } from '@providers/bankTransactions/BankTransactionsFeatureVisibility/BankTransactionsFeatureVisibilityProvider'
 import { useBankTransactionsIsCategorizationEnabledContext } from '@contexts/BankTransactionsIsCategorizationEnabledContext/BankTransactionsIsCategorizationEnabledContext'
 import { VStack } from '@ui/Stack/Stack'
 import { BankTransactionCustomerVendorSelector } from '@features/bankTransactions/BankTransactionCustomerVendorSelector/BankTransactionCustomerVendorSelector'
