+++ src/components/features/bankTransactions/BankTransactionReceipts/BankTransactionReceipts.tsx
@@ -4,8 +4,8 @@ import { useTranslation } from 'react-i18next'
 import { type BankTransaction } from '@internal-types/bankTransactions'
 import { RECEIPT_ALLOWED_INPUT_FILE_TYPES } from '@hooks/legacy/useReceipts'
 import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
-import { ReceiptsProvider } from '@providers/ReceiptsProvider/ReceiptsProvider'
-import { useReceiptsContext } from '@contexts/ReceiptsContext/ReceiptsContext'
+import { useReceiptsContext } from '@providers/bankTransactions/Receipts/ReceiptsContext'
+import { ReceiptsProvider } from '@providers/bankTransactions/Receipts/ReceiptsProvider'
 import { FileInput } from '@ui/Input/FileInput'
 import { P } from '@ui/Typography/Text'
 import { FileThumb } from '@blocks/FileThumb/FileThumb'
