+++ src/components/features/bankTransactions/BankTransactions/BankTransactions.tsx
@@ -7,17 +7,17 @@ import { usePreloadCustomAccounts } from '@api/businesses/[business-id]/custom-a
 import { usePreloadCustomers } from '@api/businesses/[business-id]/customers/get'
 import { usePreloadTagDimensions } from '@api/businesses/[business-id]/tags/dimensions/get'
 import { usePreloadVendors } from '@api/businesses/[business-id]/vendors/get'
+import { BankTransactionsProvider } from '@providers/bankTransactions/BankTransactions/BankTransactionsProvider'
+import { BankTransactionsFeatureVisibilityProvider } from '@providers/bankTransactions/BankTransactionsFeatureVisibility/BankTransactionsFeatureVisibilityProvider'
+import { BankTransactionsFiltersProvider } from '@providers/bankTransactions/BankTransactionsFiltersContext/BankTransactionsFiltersProvider'
+import { BankTransactionsPaginationProvider } from '@providers/bankTransactions/BankTransactionsPagination/BankTransactionsPaginationProvider'
+import { BankTransactionsRouteStoreProvider } from '@providers/bankTransactions/BankTransactionsRouteStore/BankTransactionsRouteStoreProvider'
+import { BankTransactionsStringOverridesProvider } from '@providers/bankTransactions/BankTransactionsStringOverridesContext/BankTransactionsStringOverridesContext'
+import { BulkSelectionStoreProvider } from '@providers/bankTransactions/BulkSelectionStore/BulkSelectionStoreProvider'
 import { BankTransactionsCategorizationStoreProvider } from '@providers/BankTransactionsCategorizationStore/BankTransactionsCategorizationStoreProvider'
-import { BankTransactionsFeatureVisibilityProvider } from '@providers/BankTransactionsFeatureVisibility/BankTransactionsFeatureVisibilityProvider'
-import { BankTransactionsProvider } from '@providers/BankTransactionsPaginationProvider/BankTransactionsProvider'
-import { BankTransactionsPaginationProvider } from '@providers/BankTransactionsProvider/BankTransactionsPaginationProvider'
-import { BankTransactionsRouteStoreProvider } from '@providers/BankTransactionsRouteStore/BankTransactionsRouteStoreProvider'
-import { BulkSelectionStoreProvider } from '@providers/BulkSelectionStore/BulkSelectionStoreProvider'
 import { InAppLinkProvider, type LinkingMetadata } from '@providers/global/InAppLink/InAppLinkContext'
 import { type BankTransactionsMode, LegacyModeProvider } from '@providers/global/LegacyModeProvider/LegacyModeProvider'
-import { BankTransactionsFiltersProvider } from '@contexts/BankTransactionsFiltersContext/BankTransactionsFiltersProvider'
 import { BankTransactionsIsCategorizationEnabledProvider } from '@contexts/BankTransactionsIsCategorizationEnabledContext/BankTransactionsIsCategorizationEnabledContext'
-import { BankTransactionsStringOverridesProvider } from '@contexts/BankTransactionsStringOverridesContext/BankTransactionsStringOverridesContext'
 import { CategorizationRulesProvider } from '@contexts/CategorizationRulesContext/CategorizationRulesContext'
 import { ErrorBoundary } from '@components/utility/ErrorBoundary/ErrorBoundary'
 import { BankTransactionsRouter } from '@features/bankTransactions/BankTransactionsRouter/BankTransactionsRouter'
