+++ src/components/features/bankTransactions/BankTransactionsBulkActions/BankTransactionsBulkActions.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
 
 import { DisplayState } from '@internal-types/bankTransactions'
-import { useBankTransactionsContext } from '@contexts/BankTransactionsContext/BankTransactionsContext'
+import { useBankTransactionsContext } from '@providers/bankTransactions/BankTransactions/BankTransactionsContext'
 import { Button } from '@ui/Button/Button'
 import { HStack } from '@ui/Stack/Stack'
 import { CategorizationMode } from '@features/bankTransactions/BankTransactionsBulkActions/BankTransactionsCategorizeAllModal'
