+++ src/components/features/bankTransactions/BankTransactionsMobileCategorySelection/BankTransactionsMobileCategorySelection.tsx
@@ -8,8 +8,8 @@ import type { TaxCodeComboBoxOption } from '@internal-types/taxCodeComboBoxOptio
 import { canCategoryHaveTaxCode } from '@utils/bankTransactions/taxCode'
 import { useGetBankTransactionCategorizationWithDefault } from '@hooks/features/bankTransactions/useGetBankTransactionCategorizationWithDefault'
 import { useTaxCodeOptions } from '@hooks/features/bankTransactions/useTaxCodeOptions'
+import { BankTransactionsFeature, useIsBankTransactionsFeatureEnabled } from '@providers/bankTransactions/BankTransactionsFeatureVisibility/BankTransactionsFeatureVisibilityProvider'
 import { useBankTransactionsCategorizationActions } from '@providers/BankTransactionsCategorizationStore/BankTransactionsCategorizationStoreProvider'
-import { BankTransactionsFeature, useIsBankTransactionsFeatureEnabled } from '@providers/BankTransactionsFeatureVisibility/BankTransactionsFeatureVisibilityProvider'
 import { VStack } from '@ui/Stack/Stack'
 import { Span } from '@ui/Typography/Text'
 import { BankTransactionTaxCodeDrawer } from '@features/bankTransactions/BankTransactionTaxCodeSelect/BankTransactionTaxCodeDrawer'
