+++ src/components/features/bankTransactions/BankTransactionsMobileList/BankTransactionsMobileList.tsx
@@ -3,10 +3,10 @@ import { useTranslation } from 'react-i18next'
 
 import { type BankTransaction } from '@internal-types/bankTransactions'
 import { useUpsertBankTransactionsDefaultCategories } from '@hooks/features/bankTransactions/useUpsertBankTransactionsDefaultCategories'
-import { useBulkSelectionActions } from '@providers/BulkSelectionStore/BulkSelectionStoreProvider'
-import { useMobileListBulkSelection } from '@providers/BulkSelectionStore/useMobileListBulkSelection'
-import { useBankTransactionsContext } from '@contexts/BankTransactionsContext/BankTransactionsContext'
-import { useBankTransactionsFiltersContext } from '@contexts/BankTransactionsFiltersContext/BankTransactionsFiltersContext'
+import { useBankTransactionsContext } from '@providers/bankTransactions/BankTransactions/BankTransactionsContext'
+import { useBankTransactionsFiltersContext } from '@providers/bankTransactions/BankTransactionsFiltersContext/BankTransactionsFiltersContext'
+import { useBulkSelectionActions } from '@providers/bankTransactions/BulkSelectionStore/BulkSelectionStoreProvider'
+import { useMobileListBulkSelection } from '@providers/bankTransactions/BulkSelectionStore/useMobileListBulkSelection'
 import { VStack } from '@ui/Stack/Stack'
 import { MobileList } from '@blocks/MobileList/MobileList'
 import { useMobileListExpansion } from '@blocks/MobileList/useMobileListExpansion'
