+++ src/components/features/bankTransactions/BankTransactionsMobileList/BankTransactionsMobileListItem.tsx
@@ -5,8 +5,8 @@ import { type BankTransaction } from '@internal-types/bankTransactions'
 import { CategorizationStatus } from '@schemas/bankTransactions/bankTransaction'
 import { convertMatchDetailsToLinkingMetadata } from '@schemas/bankTransactions/match'
 import { hasReceipts, isCategorized, isMoneyIn } from '@utils/bankTransactions/shared'
+import { useBankTransactionsContext } from '@providers/bankTransactions/BankTransactions/BankTransactionsContext'
 import { type LinkingMetadata, useInAppLinkContext } from '@providers/global/InAppLink/InAppLinkContext'
-import { useBankTransactionsContext } from '@contexts/BankTransactionsContext/BankTransactionsContext'
 import { HStack, VStack } from '@ui/Stack/Stack'
 import { Span } from '@ui/Typography/Text'
 import { BankTransactionsAmountDate } from '@features/bankTransactions/BankTransactionsAmountDate/BankTransactionsAmountDate'
