+++ src/components/features/bankTransactions/BankTransactionsRouter/BankTransactionsRouter.tsx
@@ -1,8 +1,8 @@
 import { useEffect } from 'react'
 
-import { useBankAccountFilterActions } from '@providers/BankAccountsFilterStore/BankAccountsFilterStoreProvider'
-import { BankTransactionsRoute, useBankTransactionsRouteState } from '@providers/BankTransactionsRouteStore/BankTransactionsRouteStoreProvider'
-import { useCountSelectedIds } from '@providers/BulkSelectionStore/BulkSelectionStoreProvider'
+import { useBankAccountFilterActions } from '@providers/bankTransactions/BankAccountsFilterStore/BankAccountsFilterStoreProvider'
+import { BankTransactionsRoute, useBankTransactionsRouteState } from '@providers/bankTransactions/BankTransactionsRouteStore/BankTransactionsRouteStoreProvider'
+import { useCountSelectedIds } from '@providers/bankTransactions/BulkSelectionStore/BulkSelectionStoreProvider'
 import { ResponsiveBankTransactionsView, type ResponsiveBankTransactionsViewProps } from '@features/bankTransactions/ResponsiveBankTransactionsView/ResponsiveBankTransactionsView'
 import { ResponsiveCategorizationRulesView } from '@features/categorization/ResponsiveCategorizationRulesView/ResponsiveCategorizationRulesView'
