+++ src/components/features/bankTransactions/BankTransactionsSearchField/BankTransactionsSearchField.tsx
@@ -3,8 +3,8 @@ import { useTranslation } from 'react-i18next'
 
 import { useEmitLayerEvent } from '@hooks/useEmitLayerEvent'
 import { useDebounce } from '@hooks/utils/debouncing/useDebounce'
+import { useBankTransactionsFiltersContext } from '@providers/bankTransactions/BankTransactionsFiltersContext/BankTransactionsFiltersContext'
 import { LayerEventComponent, LayerEventType } from '@providers/global/LayerProvider/layerEvents'
-import { useBankTransactionsFiltersContext } from '@contexts/BankTransactionsFiltersContext/BankTransactionsFiltersContext'
 import { SearchField } from '@ui/SearchField/SearchField'
 
 type TransactionsSearchProps = {
