+++ src/components/features/bankTransactions/ResponsiveBankTransactionsView/ResponsiveBankTransactionsView.tsx
@@ -4,9 +4,9 @@ import { BREAKPOINTS } from '@utils/screenSizeBreakpoints'
 import { unsafeAssertUnreachable } from '@utils/switch/assertUnreachable'
 import { useElementSize } from '@hooks/utils/size/useElementSize'
 import { useIsVisible } from '@hooks/utils/visibility/useIsVisible'
+import { useBankTransactionsContext } from '@providers/bankTransactions/BankTransactions/BankTransactionsContext'
+import { useBankTransactionsFiltersContext } from '@providers/bankTransactions/BankTransactionsFiltersContext/BankTransactionsFiltersContext'
 import { useBankAccountsContext } from '@contexts/BankAccountsContext/BankAccountsContext'
-import { useBankTransactionsContext } from '@contexts/BankTransactionsContext/BankTransactionsContext'
-import { useBankTransactionsFiltersContext } from '@contexts/BankTransactionsFiltersContext/BankTransactionsFiltersContext'
 import { CategorizationRulesContext } from '@contexts/CategorizationRulesContext/CategorizationRulesContext'
 import { Container } from '@blocks/Layout/Container/Container'
 import { type BankTransactionsProps } from '@features/bankTransactions/BankTransactions/BankTransactions'
