+++ src/components/features/categorization/ResponsiveCategorizationRulesView/ResponsiveCategorizationRulesView.tsx
@@ -10,7 +10,7 @@ import { useGetCategories } from '@api/businesses/[business-id]/categories/get'
 import { usePostArchiveCategorizationRule } from '@api/businesses/[business-id]/categorization-rules/[categorization-rule-id]/archive/post'
 import { useDebouncedSearchProps } from '@hooks/utils/debouncing/useDebouncedSearchQuery'
 import { useSizeClass } from '@hooks/utils/size/useWindowSize'
-import { useBankTransactionsNavigation, useCategorizationRulesTableFilters } from '@providers/BankTransactionsRouteStore/BankTransactionsRouteStoreProvider'
+import { useBankTransactionsNavigation, useCategorizationRulesTableFilters } from '@providers/bankTransactions/BankTransactionsRouteStore/BankTransactionsRouteStoreProvider'
 import { useLayerContext } from '@providers/global/LayerContext/LayerContext'
 import { Button } from '@ui/Button/Button'
 import { DataState, DataStateStatus } from '@ui/DataState/DataState'
