+++ src/hooks/features/bankTransactions/useAugmentedBankTransactions.tsx
@@ -10,7 +10,7 @@ import { type BankTransactionFilters } from '@utils/bankTransactions/shared'
 import { type UseBankTransactionsOptions, useGetListBankTransactions } from '@api/businesses/[business-id]/bank-transactions/get'
 import { useFilterBankTransactions } from '@hooks/features/bankTransactions/useFilterBankTransactions'
 import { usePollBankTransactions } from '@hooks/features/bankTransactions/usePollBankTransactions'
-import { useBankTransactionsFiltersContext } from '@contexts/BankTransactionsFiltersContext/BankTransactionsFiltersContext'
+import { useBankTransactionsFiltersContext } from '@providers/bankTransactions/BankTransactionsFiltersContext/BankTransactionsFiltersContext'
 import { CategorizationRulesContext } from '@contexts/CategorizationRulesContext/CategorizationRulesContext'
 
 const tagFilterToParams = (tagFilter: TagFilterInput): Pick<UseBankTransactionsOptions, 'tagKey' | 'tagValues'> => {
