+++ src/hooks/features/bankTransactions/useBankTransactionsPagination.ts
@@ -5,7 +5,7 @@ import { type BankTransactionFilters } from '@utils/bankTransactions/shared'
 import { useEmitLayerEvent } from '@hooks/useEmitLayerEvent'
 import { PaginationChangeSource, type TablePaginationProps } from '@hooks/utils/pagination/types'
 import { useTablePaginationProps } from '@hooks/utils/pagination/useTablePaginationProps'
-import { useCurrentBankTransactionsPage } from '@providers/BankTransactionsRouteStore/BankTransactionsRouteStoreProvider'
+import { useCurrentBankTransactionsPage } from '@providers/bankTransactions/BankTransactionsRouteStore/BankTransactionsRouteStoreProvider'
 import { LayerEventComponent, LayerEventType } from '@providers/global/LayerProvider/layerEvents'
 
 type UseBankTransactionsPaginationParams = {
