+++ src/hooks/features/bankTransactions/useCategorizeBankTransactionWithCacheUpdate.ts
@@ -5,8 +5,8 @@ import type { CategoryUpdate } from '@schemas/bankTransactions/categoryUpdate'
 import { usePutCategorizeBankTransaction } from '@api/businesses/[business-id]/bank-transactions/[bank-transaction-id]/categorize/put'
 import { useBankTransactionsGlobalCacheActions } from '@api/businesses/[business-id]/bank-transactions/get'
 import { useProfitAndLossGlobalInvalidator } from '@api/businesses/[business-id]/reports/profit-and-loss/useProfitAndLossGlobalInvalidator'
+import { useBankTransactionsContext } from '@providers/bankTransactions/BankTransactions/BankTransactionsContext'
 import { useLayerContext } from '@providers/global/LayerContext/LayerContext'
-import { useBankTransactionsContext } from '@contexts/BankTransactionsContext/BankTransactionsContext'
 
 export function useCategorizeBankTransactionWithCacheUpdate() {
   const { eventCallbacks } = useLayerContext()
