+++ src/hooks/features/bankTransactions/useIsEditableCustomBankTransaction.ts
@@ -1,7 +1,7 @@
 import type { BankTransaction } from '@internal-types/bankTransactions'
 import { isEditableCustomTransaction } from '@utils/bankTransactions/shared'
 import { useGetCustomAccounts } from '@api/businesses/[business-id]/custom-accounts/get'
-import { BankTransactionsFeature, useIsBankTransactionsFeatureEnabled } from '@providers/BankTransactionsFeatureVisibility/BankTransactionsFeatureVisibilityProvider'
+import { BankTransactionsFeature, useIsBankTransactionsFeatureEnabled } from '@providers/bankTransactions/BankTransactionsFeatureVisibility/BankTransactionsFeatureVisibilityProvider'
 import { useBankTransactionsIsCategorizationEnabledContext } from '@contexts/BankTransactionsIsCategorizationEnabledContext/BankTransactionsIsCategorizationEnabledContext'
 
 export const useIsEditableCustomBankTransaction = (bankTransaction: BankTransaction) => {
