+++ src/hooks/legacy/useReceipts.ts
@@ -11,8 +11,8 @@ import { hasReceipts } from '@utils/bankTransactions/shared'
 import { useEmitLayerEvent } from '@hooks/useEmitLayerEvent'
 import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
 import { useBuildKeyInputs } from '@hooks/utils/swr/useBuildKeyInputs'
+import { useBankTransactionsContext } from '@providers/bankTransactions/BankTransactions/BankTransactionsContext'
 import { LayerEventComponent, LayerEventType } from '@providers/global/LayerProvider/layerEvents'
-import { useBankTransactionsContext } from '@contexts/BankTransactionsContext/BankTransactionsContext'
 
 const listBankTransactionDocuments = get<{
   data: DocumentS3Urls
