+++ src/test-utils/bankTransactionsStoryControls.tsx
@@ -1,6 +1,6 @@
 import { type Meta } from '@storybook/react-vite'
 
-import { DEFAULT_FEATURE_VISIBILITY } from '@providers/BankTransactionsFeatureVisibility/BankTransactionsFeatureVisibilityProvider'
+import { DEFAULT_FEATURE_VISIBILITY } from '@providers/bankTransactions/BankTransactionsFeatureVisibility/BankTransactionsFeatureVisibilityProvider'
 import { type MobileComponentType } from '@features/bankTransactions/constants'
 
 // The bank-transactions props shared by every component that embeds the
